roxlap-gpu + roxlap-render build and run on wasm32 (WebGPU). roxlap-web and roxlap-cave-web render through the SceneRenderer facade: the wgpu compute marcher when WebGPU is available, else CPU opticast presented via a WebGL2 blit owned by the facade.
  - roxlap-gpu/render compile for wasm32: async canvas constructors create the
    surface via SurfaceTarget::Canvas; pollster/softbuffer/headless gated
    native-only; the wasm canvas path drops the Send+Sync bound; wasm pick_depth
    returns None.
  - CPU fallback presents on the web via a WebGL2 blit (cpu_blit.rs, unsafe-free
    via bytemuck) instead of softbuffer.
  - wasm GPU init probes adapter AND device before create_surface, so a failed
    init leaves the canvas pristine for the WebGL2 fallback.
  - Upgrade wgpu 22->29 (+ naga 29, egui/egui-wgpu 0.34): wgpu 22 sent the
    spec-removed maxInterStageShaderComponents limit, which modern Chrome
    rejects, breaking WebGPU device creation.
  - Grid::bake_lightmode: reusable neighbour-aware estnorm lighting bake.
  - roxlap-web: procedural terraced-hills Scene. roxlap-cave-web: single-chunk
    cave Scene with scene collision/carving and bullets as facade sprites.
  - Bumps the workspace to 1.0.0.

+++ crates/roxlap-gpu/src/scene.rs
@@ -825,7 +825,8 @@ fn split_occupancy_pages(
     slot_align_words: u64,
 ) -> (Vec<wgpu::Buffer>, u32, u32) {
     let total_words = words.len() as u64;
-    let limit_words = u64::from(device.limits().max_storage_buffer_binding_size) / 4;
+    // wgpu 29 widened `max_storage_buffer_binding_size` to `u64`.
+    let limit_words = device.limits().max_storage_buffer_binding_size / 4;
     // Largest slot-aligned page that fits one binding (≥ 1 slot).
     let page_slots = (limit_words / slot_align_words).max(1);
     let mut page_words = page_slots.saturating_mul(slot_align_words);
